+++ src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod file_io;
 pub mod config;
-pub mod day;
 pub mod work_summary;
